// sdk/data/test/database.types.ts
// Generated by gregale data-api types. Do not edit.
// Schema fingerprint: 1c1609083e51ef6e760d6431bd672e27ffb61c6fe47206069a876bc6a41dbdaa
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[]
export type Database = {
  "api": {
    Tables: {
      "comments": {
        Row: {
          "id": number
          "note_id": number
          "body": string
        }
        Insert: {
          "id": number
          "note_id": number
          "body": string
        }
        Update: {
          "id"?: number
          "note_id"?: number
          "body"?: string
        }
        Relationships: [{"foreignKeyName":"comments_note_id_fkey","columns":["note_id"],"isOneToOne":false,"referencedRelation":"notes","referencedColumns":["id"]}]
      }
      "notes": {
        Row: {
          "id": number
          "subject": string
          "body": string
          "state": "open" | "closed" | null
          "tags": (string | null)[]
          "payload": Json | null
          "length": number | null
        }
        Insert: {
          "id"?: never
          "subject": string
          "body": string
          "state"?: "open" | "closed" | null
          "tags"?: (string | null)[]
          "payload"?: Json | null
          "length"?: never
        }
        Update: {
          "id"?: never
          "subject"?: string
          "body"?: string
          "state"?: "open" | "closed" | null
          "tags"?: (string | null)[]
          "payload"?: Json | null
          "length"?: never
        }
        Relationships: []
      }
    }
    Views: {
      "note_bodies": {
        Row: {
          "id": number | null
          "body": string | null
        }
        Relationships: []
      }
    }
    Functions: {
    }
    Enums: {
      "state": "open" | "closed"
    }
    CompositeTypes: Record<string, never>
  }
}